Rename cleansing-flame-source-type.h to cflame-source.h
So that the file's name remains the same as that of the enum (class).
