+++ crawl-ref/source/cflame-source.h
